TL: main coupling done, need to adapt to new dedalus version
